Merged PR 202609: MSDateTimePicker with MSDatePicker integration
This PR adds a public `MSDateTimePicker` class to display and select a date (with optional time) from a `UIPickerView` style view. It has public APIs to programmatically change the selected date, present the modal, and set a delegate that will be notified when the user has confirmed a date.

It can be presented directly, but is also used as a replacement for `MSDatePicker` when Voiceover is turned on, so that the user can interact with the various date components.

![Screen Shot 2018-11-13 at 6.39.11 PM.png](https://onedrive.visualstudio.com/4dcbf0bc-c3cd-49c8-a7c3-ec1924691d9b/_apis/git/repositories/93ac71ee-b53a-4fc6-a8c4-d46a80d4ca39/pullRequests/202609/attachments/Screen%20Shot%202018-11-13%20at%206.39.11%20PM.png)

Related work items: #655982