Merged PR 435213: [NuGet] Update NuGet Publish to Xcode 11.4.1
Office and the new BuildXL build system require our builds to happen with 11.4.1 to prevent a write access failure with how Swift handles libraries built on older versions of the compiler.

Update to a public 10.15 image that has 11.4.1 as its default Xcode version.

Note: Don't invest right now into changing xcode version from "Default" as this repo is deprecated and soon to be replaced.